Introduce Tuttifrutti.Cache.Storage
Tuttifrutti.Cache.Storage provides an interface intended for storing
cached values with a basic set of operations to CRUD and expire the
items.

The provided implemnetation (Tuttifrutti.Cache.Storage.InMemory)
maintains the cache in an in-memory HashPSQ. A postgres backend is
coming soon.

All of that is based on cache implementations in Faro and Persona
and should cover the needs of both.