Add InMemoryRef cache storage backend
The only difference from `InMemory` is that this one doesn't serialize
keys/values to the bytestrings and keeps track of number of entries
not number of bytes.

The advantage is that it allows to cache arbitrary value references
without serialization/deserialization overhead. The disadvantage is
that it's not possible to set a precise memory limit (only max amount
of entries).

The code is essentially copied and unneeded parts are stripped (removed the
bit-counting things, and serialization/deserialization).